add ep-eth, use it from pallet
