+++ frame/ethink/Cargo.toml
@@ -35,6 +35,7 @@ pallet-transaction-payment = { workspace = true, features = ["default"] }
 pallet-insecure-randomness-collective-flip = { workspace = true, features = ["default"] }
 ep-crypto = { workspace = true, features = ["default"] }
 ep-mapping = { workspace = true, features = ["default"] }
+ep-eth = { workspace = true, features = ["default"] }
 sp-core = { workspace = true, features = ["default"] }
 
 [features]
