+++ primitives/crypto/Cargo.toml
@@ -3,6 +3,7 @@ name = "ep-crypto"
 version = "0.1.0"
 license = "Apache-2.0"
 description = "Crypto primitives for ethink!"
+authors.workspace = true
 edition.workspace = true
 
 [dependencies]
