+++ primitives/eth/src/lib.rs
@@ -1,27 +1,29 @@
-// SPDX-License-Identifier: Apache-2.0
-//
-// This file is part of Ethink.
-//
-// Copyright (c) 2023-2024 Alexander Gryaznov.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// 	http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+use ep_crypto::EthereumSignature;
+use ep_mapping::{SubstrateWeight, Weight};
+use sp_core::{U256, ecdsa, Pair};
 
-use crate::{
-    common::contracts::ContractInput, ecdsa, EthTransaction, EthereumSignature, LegacyTransaction,
-    LegacyTransactionMessage, SubstrateWeight, TransactionSignature, Weight, U256,
+pub use ethereum::{
+    TransactionV2 as EthTransaction,
+    LegacyTransaction,
+    LegacyTransactionMessage,
 };
 
-use sp_core::crypto::Pair;
+// TODO: move ethereum-types re-exports here
+
+#[derive(Clone)]
+pub struct ContractInput(Vec<u8>);
+
+impl From<Vec<u8>> for ContractInput {
+    fn from(v: Vec<u8>) -> Self {
+        Self(v)
+    }
+}
+
+impl Into<Vec<u8>> for ContractInput {
+    fn into(self) -> Vec<u8> {
+        self.0
+    }
+}
 
 #[derive(Clone)]
 /// Ethereum transaction input, used for transaciton building in tests
@@ -74,7 +76,7 @@ impl From<EthTxInput> for LegacyTransactionMessage {
 pub fn compose_and_sign_tx(i: EthTxInput) -> EthTransaction {
     let msg: LegacyTransactionMessage = i.clone().into();
     let sig = EthereumSignature::new(i.signer.sign_prehashed(&msg.hash().into()));
-    let sig: Option<TransactionSignature> = sig.into();
+    let sig: Option<ethereum::TransactionSignature> = sig.into();
     let signature = sig.expect("signer generated no signature");
 
     EthTransaction::Legacy(LegacyTransaction {
